This is release 0.74 of es.  This is a not quite feature complete release, but it's getting close; see the file TODO for details on what remains to be done.  We hope to release 0.8 within the next 10 days.  See the file CHANGES for what's been changed since the last distribution.
This release has been tested (though not very thoroughly) on the
following machine/compiler combinations:

	NeXT OS2.1, gcc 1.36
	Sparc, SunOS 4.1.2, gcc 1.39
	Sun3, SunOS 4.0.3, gcc 1.37.1
	IBM RS/6000, AIX 3.2, native compiler
	Digital Alpha AXP, OSF/1 T1.2-2, native compiler
	DECstation 5000, Ultrix, gcc 1.39

Please tell us of any other machines which es works correctly on.

One bug to be aware of is that the interaction between control-c
and the new interpreter loops is somewhat buggy.  After you hit a
control-c, you may want to type return a couple of times.  This
will be fixed in the next release.