job control
I wrote a patch which adds job control to es similar to those found in the bourne shell derivatives. Your system must have the waitpid subroutine, as the patch replaces the calls to the wait3 subroutine by a call to waitpid. Therefore the builtin time may return sensless values, dependening on the definition of the HAVE_WAIT3 macro.