fixing es builds but does not run on amd64/FBSD 6.
mksignal tries to build some definitions, but at least for my system adds
too many and runs afoul of the check in signals.c
In particular, SIGLWP == NSIG

Note:
/sys/signal.h:#define SIGLWP SIGTHR
and SIGTHR was already a problem.

I just added 2 lines
that do a check against NSIG when defining signals. This seems better to
me than just adding ignores as it's general -- it will still work if the
signals get renumbered (as they apparently were).