Fix segfault in match
These either abort in assertion failure at match.c:149 or segfault:

  es -c '~ 0 1 `{}`{}`{}`{}`{} 0'
  es -c '~ `{} 1 `{}`{}`{}`{}`{} 0'
  es -c '~~ 0 1 `{}`{}`{}`{}`{}`{}`{}`{}`{} 0'

Functions listmatch and extractmatches require their pattern and quote
arguments to be lists of the same length.  These lists are prepared in
matchpattern and extractpattern using glom2.  When GC is performed
inside glom2, the head of the quote list would be freed.  This would
have no observable effects unless more items are added afterwards, in
which case quote would point to the shorter list in the old unforwarded
(freed) space violating the assumption about its length.

Add two assertions to catch length mismatches before segfault, and add
quote list to rootlist while it is in use so that it gets properly
forwarded on GC.