Improve GNU Readline support in configure
GNU Readline is now the only supported readline implementation, so any
mention of editline is gone.

readline.m4 defines a macro ES_CHECK_READLINE that performs a linker
test to ensure that using readline headers and libraries will work,
checking which of the following work:

* `-lreadline`
* `-lreadline -lcurses`
* `-lreadline -ltinfo`
* `-lreadline -lterminfo`
* `-lreadline -ltermcap`

readline.m4 also defines a macro ES_WITH_READLINE that adds a
`--with-readline` configure flag that defaults to a value of "check" to
conditionally enable readline support by default when possible.
The ES_WITH_READLINE macro additionally defines two "precious" variables
READLINE_CFLAGS and READLINE_LIBS, used as the CFLAGS and LIBS arguments
to ES_CHECK_READLINE if set.
If `--with-readline[=yes]` was specified and all attempts at
compilation/linking fail, the default behavior is to terminate configure
with a failure message.
Otherwise, readline functionality is silently disabled, even if
READLINE_CFLAGS and/or READLINE_LIBS were specified.