Add hook to control filename completion.
The new %completion-to-file function, when defined within %complete,
controls whether and how readline treats completion candidates as
file names.  This formats completion menus much prettier, only showing
the basename of files instead of the full path, and also enables much
nicer behavior with directory completion.

A very nice concrete example is a function `in`, which takes a directory
and a command as arguments and runs the command in the directory.  With
a custom %completion-to-file, the completion of the "internal" command
can be done wholly as if it's already being invoked within the
directory.