Add a few scripts to examples/.
These are of the same "form" as the canonical extensions, such that
sourcing them from .esrc will modify the shell in some more-or-less
useful way.  However, they don't have the same historical pedigree or
universally acknowledged usefulness as the canonical extensions, so
instead we put them in examples/ until people decide they should be
installed with every copy of es.