Move 'retry' exception handling from $&catch to fn-catch
With this change, the $&catch primitive is about as complex as a
primitive should be.  Arguably, 'retry' should be removed from the shell
entirely, but this is a backwards-compatible half-measure, and
it does demonstrate how to "reimplement" retry in case it is actually
removed and people want to add it back.